# src/main/java/com/vnm/creditcardprocessor/model/CCError.java
/**
 *
 */
package com.vnm.creditcardprocessor.model;
/**
 * @author  varmathu0
 * @project Credit-Card-Processor
 */
public class CCError {

    private String message;

    public String getMessage() {
        return message;
    }

    public void setMessage(String message) {
        this.message = message;
    }

}